move and rename plot_signature to plot_filter
